+++ src/main/java/com/example/yamyam16/menu/dto/MenuUpdateResponseDto.java
@@ -11,6 +11,6 @@ public class MenuUpdateResponseDto {
 
 	private String menuName;
 
-	private int menuPrice;
+	private long menuPrice;
 
 }
